move constants to root directory to avoid loading of external modules in setup via panoply.__init__
