+++ setup.py
@@ -1,5 +1,6 @@
 from distutils.core import setup
-from panoply import VERSION, PKGNAME
+from constants import VERSION, PKGNAME
+
 setup(
     name=PKGNAME,
     version=VERSION,
